+++ app/tasks/probe/LowRateReponseConsoTask.scala
@@ -7,16 +7,16 @@ import play.api.Logger
 import repositories.probe.ProbeRepository
 import repositories.tasklock.TaskRepositoryInterface
 import repositories.user.UserRepositoryInterface
-import services.emails.Email.AdminProbeTriggered
+import services.emails.EmailDefinitionsAdmin.AdminProbeTriggered
 import services.emails.MailService
 import tasks.ScheduledTask
 import utils.Logs.RichLogger
 
 import java.time.LocalTime
-import scala.concurrent.duration.DurationInt
-import scala.concurrent.duration.FiniteDuration
 import scala.concurrent.ExecutionContext
 import scala.concurrent.Future
+import scala.concurrent.duration.DurationInt
+import scala.concurrent.duration.FiniteDuration
 
 class LowRateReponseConsoTask(
     actorSystem: ActorSystem,
@@ -38,7 +38,9 @@ class LowRateReponseConsoTask(
       for {
         users <- userRepository.listForRoles(Seq(UserRole.Admin))
         _ <- mailService
-          .send(AdminProbeTriggered(users.map(_.email), "Taux de signalements 'Réponse conso' faible", rate, "bas"))
+          .send(
+            AdminProbeTriggered.build(users.map(_.email), "Taux de signalements 'Réponse conso' faible", rate, "bas")
+          )
       } yield ()
     case rate =>
       logger.debug(s"Taux de signalements correct: $rate%")
